+++ tests/unit/test_shelf.py
@@ -22,6 +22,15 @@ def shelf(local_bookshelf):
     return shelf
 
 
+def get_meta(conn, bucket, pkg):
+    volume_meta_contents = io.BytesIO()
+    conn.Object(bucket, f"/this/prefix/{pkg}/volume.json").download_fileobj(
+        volume_meta_contents
+    )
+    volume_meta_contents.seek(0)
+    return json.load(volume_meta_contents)
+
+
 def setup_upload_bucket(remote_bookshelf, monkeypatch):
     remote_bookshelf.mocker.get(
         f"/{DATA_FORMAT_VERSION}/new-package/volume.json", status_code=404
@@ -126,13 +135,7 @@ def test_publish(shelf, remote_bookshelf, monkeypatch, caplog, example_data):
     )
     conn.Object(bucket, "/this/prefix/new-package/v1.1.1_e001/example.csv").load()
 
-    volume_meta_contents = io.BytesIO()
-    conn.Object(bucket, "/this/prefix/new-package/volume.json").download_fileobj(
-        volume_meta_contents
-    )
-    volume_meta_contents.seek(0)
-    volume_meta = json.load(volume_meta_contents)
-
+    volume_meta = get_meta(conn, bucket, "new-package")
     assert len(volume_meta["versions"]) == 1
     assert volume_meta["versions"][-1]["version"] == "v1.1.1"
 
@@ -154,17 +157,37 @@ def test_publish_new_version(
     bucket = os.environ["BOOKSHELF_BUCKET"]
     conn.Object(bucket, "/this/prefix/test/v1.1.1_e025/datapackage.json").load()
 
-    volume_meta_contents = io.BytesIO()
-    conn.Object(bucket, "/this/prefix/test/volume.json").download_fileobj(
-        volume_meta_contents
-    )
-    volume_meta_contents.seek(0)
-    volume_meta = json.load(volume_meta_contents)
-
+    volume_meta = get_meta(conn, bucket, "test")
     assert len(volume_meta["versions"]) == 3  # 2 existing in volume.json + one extra
     assert volume_meta["versions"][-1]["version"] == "v1.1.1"
 
     assert "Book test@v1.1.1 ed.25 uploaded successfully" in caplog.text
+
+
+@pytest.mark.parametrize("edition", [2, 900])
+@moto.mock_s3
+def test_publish_new_edition(
+    shelf, remote_bookshelf, monkeypatch, caplog, example_data, edition
+):
+    conn = setup_upload_bucket(remote_bookshelf, monkeypatch)
+    book = LocalBook.create_new("test", "v1.1.0", edition=edition)
+
+    with caplog.at_level(logging.INFO):
+        shelf.publish(book)
+
+    # Check that files uploaded
+    bucket = os.environ["BOOKSHELF_BUCKET"]
+    conn.Object(
+        bucket, f"/this/prefix/test/v1.1.0_e{edition:03}/datapackage.json"
+    ).load()
+
+    volume_meta = get_meta(conn, bucket, "test")
+    assert len(volume_meta["versions"]) == 3  # 2 existing in volume.json + one extra
+    assert volume_meta["versions"][-1]["version"] == "v1.1.0"
+    assert volume_meta["versions"][-1]["edition"] == edition
+
+    assert "Uploading a new edition of an existing book" in caplog.text
+    assert f"Book test@v1.1.0 ed.{edition} uploaded successfully" in caplog.text
 
 
 @moto.mock_s3
@@ -188,7 +211,12 @@ def test_publish_wrong_permissions(shelf, remote_bookshelf, monkeypatch, caplog)
 def test_publish_existing(shelf, remote_bookshelf):
     book = LocalBook.create_new("test", "v1.0.0", 1)
 
-    with pytest.raises(UploadError, match="Book with the same version already exists"):
+    with pytest.raises(
+        UploadError,
+        match=re.escape(
+            "Edition value has not been increased (remote: v1.0.0_e001, local: v1.0.0_e001)"
+        ),
+    ):
         shelf.publish(book, force=False)