+++ CHANGELOG.rst
@@ -19,6 +19,7 @@ Changed
 =======
 
 - (`!19 <https://gitlab.com/climate-resource/bookshelf/bookshelf/merge_requests/19>`_) Added gwp_context field to primap-hist for easier post processing
+- (`!19 <https://gitlab.com/climate-resource/bookshelf/bookshelf/merge_requests/19>`_) Fixed the uploading of new editions
 
 v0.2.0
 ------